MLPerf-based-workloads/retinanet/requirements.txt
flask
waitress
numpy
pillow
onnxruntime-gpu
